moved add function to new add view. made helper function for adding a new joke. adding a joke is currently functional...can add joke to a single category.
