+++ jt/views/__init__.py
@@ -11,10 +11,12 @@ from .joke_views import random_joke
 from .joke_views import list_by_category
 from .joke_views import add_to_favorites
 from .joke_views import random_joke
-from .joke_views import add_joke
 
 from .favorite_views import favorites_list
 from .favorite_views import favorites_train
 from .favorite_views import delete_from_favorites
 
 from .search_bar_views import search
+
+from .add_new_joke import add_joke
+from .add_new_joke import add_joke_category
