+++ jt/views/joke_views.py
@@ -61,25 +61,3 @@ def random_joke(request):
   context = { 'joke_at_random' : joke_at_random }
   print('RANDOMJOKE', context)
   return render (request, 'index.html', context)
-
-
-@login_required
-def add_joke(request):
-  '''Handles user adding a new joke to the database'''
-
-  if request.method == "GET":
-    newjoke_form = NewJokeForm()
-    template_name = 'new_joke.html'
-    return render(request, template_name, {'newjoke_form': newjoke_form})
-
-  elif request.method == "POST":
-    creator = request.user
-    question = request.POST["question"]
-    answer = request.POST["answer"]
-    hint = request.POST["hint"]
-    category = request.POST["category"]
-    new_joke = Joke(creator=creator, question=question, answer=answer, hint=hint, category=category)
-    new_joke.save()
-    print('NEW JOKE ADDED?', new_joke.id)
-
-    return HttpResponseRedirect(reverse('jt:random_joke'))
